Moved ValueContainer to separate file
