+++ CMakeLists.txt
@@ -14,6 +14,8 @@ add_library(CLArgs INTERFACE
         include/CLArgs/from_string.hpp
         include/CLArgs/flag.hpp
         include/CLArgs/option.hpp
+        include/CLArgs/string_literal.hpp
+        include/CLArgs/value_container.hpp
 )
 
 target_include_directories(CLArgs INTERFACE
