+++ include/CLArgs/option.hpp
@@ -1,7 +1,7 @@
 #ifndef CLARGS_OPTION_HPP
 #define CLARGS_OPTION_HPP
 
-#include <CLArgs/core.hpp>
+#include <CLArgs/string_literal.hpp>
 
 #include <concepts>
